+++ Hook/controllers/__init__.py
@@ -1,0 +1,3 @@
+"""
+    This package should have import only in the direction of models
+"""
