+++ Hook/routes/api/user.py
@@ -1,15 +1,13 @@
 import json
-
 from flask import jsonify, g, request, Response
-from flask.ext.login import login_required
 
+import Hook.controllers.github
 from Hook.app import app
 
-@app.route("/api/rest/v1.0/user/repositories", methods=["GET", "POST"])
-@login_required
-def api_user_repositories():
 
-    if g.user:
+@app.route("/api/rest/v1.0/user/repositories", methods=["GET", "POST"])
+def api_user_repositories():
+    if hasattr(g, "user") and g.user:
         if request.method == "GET":
             response = json.dumps([repo.dict() for repo in g.user.repositories])
         elif request.method == "POST":
@@ -20,10 +18,10 @@ def api_user_repositories():
     else:
         return jsonify(None)
 
+
 @app.route("/api/rest/v1.0/user/repositories/<owner>/<name>", methods=["PUT"])
-@login_required
 def api_user_repository_switch(owner, name):
-    if g.user:
-        return jsonify(status=g.user.switch(owner, name))
+    if hasattr(g, "user") and g.user:
+        return jsonify(status=g.user.switch(owner, name, Hook.controllers.github.hook))
     else:
         return jsonify(status=False)
